+++ core/src/main/java/com/wxsoft/fcare/core/data/entity/drug/DrugPackageItem.kt
@@ -35,7 +35,7 @@ data class DrugPackageItem (val id:String): BaseObservable() {
         }
 
     @Bindable
-    var dose: Int = 0
+    var dose: Float = 0f
         set(value) {
             field = value
             notifyPropertyChanged(BR.dose)
